Simplified check passes
Earlier, I tried to make the update smart and work out what to
check based on what changed. But I never got that working. There
was a remnant of that, the checks would first check their
dependancies, then after they would check their dependants. That
was causing trouble and it wasn't necessary because of the way
dependancies are resolved.